move test resource from conflicting directory
